VP-1677 Extract logic from AddUserModal.vue to ts
